# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "API автоматизации рабочих процессов",
        "contact": {},
        "version": "1"
    },
    "host": "localhost:8080/",
    "paths": {
...
        "/management/deployments": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "deployments"
                ],
                "summary": "Получить список всех деплоев",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Response-array_deployment_handler_DeploymentResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "deployments"
                ],
                "summary": "Создать новый деплой",
                "parameters": [
                    {
                        "description": "Данные деплоя",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/deployment_handler.CreateDeploymentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Response-deployment_handler_DeploymentResponse"
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    }
                }
            }
        },
        "/management/deployments/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "deployments"
                ],
                "summary": "Получить информацию о деплое",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID деплоя",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Response-deployment_handler_DeploymentResponse"
                        }
                    },
                    "404": {
                        "description": "Деплой не найден",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации ID",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "deployments"
                ],
                "summary": "Обновить информацию о деплое",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID деплоя",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Новые данные деплоя",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/deployment_handler.UpdateDeploymentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Response-deployment_handler_DeploymentResponse"
                        }
                    },
                    "404": {
                        "description": "Деплой не найден",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "deployments"
                ],
                "summary": "Удалить деплой",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID деплоя",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Деплой не найден",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации ID",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.Error": {
            "description": "Модель ошибки на запрос",
            "type": "object",
            "properties": {
                "message": {
                    "description": "Текст ошибки",
                    "type": "string"
                }
            }
        },
        "api.Response-any": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "Данные"
                },
                "error": {
                    "description": "Ошибка",
                    "allOf": [
                        {
                            "$ref": "#/definitions/api.Error"
                        }
                    ]
                }
            }
        },
...
        "api.Response-array_deployment_handler_DeploymentResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "Данные",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/deployment_handler.DeploymentResponse"
                    }
                },
                "error": {
                    "description": "Ошибка",
                    "allOf": [
                        {
                            "$ref": "#/definitions/api.Error"
                        }
                    ]
                }
            }
        },
...
        "api.Response-deployment_handler_DeploymentResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "Данные",
                    "allOf": [
                        {
                            "$ref": "#/definitions/deployment_handler.DeploymentResponse"
                        }
                    ]
                },
                "error": {
                    "description": "Ошибка",
                    "allOf": [
                        {
                            "$ref": "#/definitions/api.Error"
                        }
                    ]
                }
            }
        },
...
        "deployment_handler.CreateDeploymentRequest": {
            "description": "Запрос на создание деплоя",
            "type": "object",
            "required": [
                "pull_request_number",
                "status",
                "thread_ts"
            ],
            "properties": {
                "pull_request_number": {
                    "description": "@Description\tНомер Pull Request\n\t@Example:\t\t42\n\t@Required\t\ttrue",
                    "type": "string"
                },
                "status": {
                    "description": "@Description\tСтатус деплоя\n\t@Example:\t\tpending\n\t@Enum\t\t\tpending,success,failed,running\n\t@Required\t\ttrue",
                    "type": "string",
                    "enum": [
                        "pending",
                        "success",
                        "failed",
                        "running"
                    ]
                },
                "thread_ts": {
                    "description": "@Description\tИдентификатор треда в Slack\n\t@Example:\t\t1234567890.123456\n\t@Required\t\ttrue",
                    "type": "string"
                },
                "workflow_run_id": {
                    "description": "@Description\tID Workflow Run в GitHub\n\t@Example:\t\t987654321",
                    "type": "string"
                }
            }
        },
        "deployment_handler.DeploymentResponse": {
            "description": "Информация о процессе деплоя",
            "type": "object",
            "properties": {
                "id": {
                    "description": "@Description\tУникальный ID деплоя\n\t@Example:\t\t1",
                    "type": "integer"
                },
                "pull_request_number": {
                    "description": "@Description\tНомер Pull Request\n\t@Example:\t\t42",
                    "type": "string"
                },
                "status": {
                    "description": "@Description\tСтатус деплоя\n\t@Example:\t\tsuccess\n\t@Enum\t\t\tpending,success,failed,running",
                    "type": "string"
                },
                "thread_ts": {
                    "description": "@Description\tИдентификатор треда в Slack\n\t@Example:\t\t1234567890.123456",
                    "type": "string"
                },
                "workflow_run_id": {
                    "description": "@Description\tID Workflow Run в GitHub\n\t@Example:\t\t987654321",
                    "type": "string"
                }
            }
        },
        "deployment_handler.UpdateDeploymentRequest": {
            "description": "Запрос на обновление деплоя",
            "type": "object",
            "properties": {
                "pull_request_number": {
                    "description": "@Description\tНомер Pull Request\n\t@Example:\t\t42",
                    "type": "string"
                },
                "status": {
                    "description": "@Description\tСтатус деплоя\n\t@Example:\t\tsuccess\n\t@Enum\t\t\tpending,success,failed,running",
                    "type": "string",
                    "enum": [
                        "pending",
                        "success",
                        "failed",
                        "running"
                    ]
                },
                "thread_ts": {
                    "description": "@Description\tИдентификатор треда в Slack\n\t@Example:\t\t1234567890.123456",
                    "type": "string"
                },
                "workflow_run_id": {
                    "description": "@Description\tID Workflow Run в GitHub\n\t@Example:\t\t987654321",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}